Separate ChildNumber
