+++ crypto/src/key/hdkd/derivation_path.rs
@@ -13,104 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::child_number::ChildNumber;
 use super::derivable::DerivationError;
-use super::u31::U31;
 use core::fmt;
-use std::fmt::{Formatter, Write};
+use std::fmt::Formatter;
 use std::str::FromStr;
 
 const PREFIX: &str = "m";
 const SEPARATOR: &str = "/";
-const HARDENED_APOS: char = '\'';
-const HARDENED_H: char = 'h';
-
-/// BIP32-like child numbers. Currently we only support hardened derivations
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Ord, PartialOrd)]
-pub struct ChildNumber(DerivationType);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Ord, PartialOrd)]
-enum DerivationType {
-    Hardened(U31),
-}
-
-impl ChildNumber {
-    /// Return a hardened child number
-    pub fn from_hardened(index: U31) -> Result<Self, DerivationError> {
-        Ok(ChildNumber(DerivationType::Hardened(index)))
-    }
-
-    /// Return a normal child number
-    fn from_normal(_index: U31) -> Result<Self, DerivationError> {
-        // For the time being we don't support non-hardened derivations
-        Err(DerivationError::UnsupportedDerivationType)
-    }
-
-    /// Return a child based on the hardened bit (MSB bit)
-    pub fn from_index_with_hardened_bit(index: u32) -> Result<Self, DerivationError> {
-        let (index_u31, msb) = U31::from_u32_with_msb(index);
-        if msb {
-            Self::from_hardened(index_u31)
-        } else {
-            Self::from_normal(index_u31)
-        }
-    }
-
-    /// Return a BIP32-like child number index that has the hardened bit set if needed
-    pub fn to_encoded_index(self) -> u32 {
-        match self.0 {
-            DerivationType::Hardened(i) => i.into_encoded_with_msb(true),
-        }
-    }
-
-    /// Returns true if this child number is hardened
-    pub fn is_hardened(&self) -> bool {
-        match self.0 {
-            DerivationType::Hardened(_) => true,
-        }
-    }
-}
-
-impl FromStr for ChildNumber {
-    type Err = DerivationError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // Check if this child number is hardened and get the index number string
-        let (index_str, is_hardened) = match s.strip_suffix([HARDENED_APOS, HARDENED_H]) {
-            Some(prefix) => (prefix, true),
-            None => (s, false),
-        };
-
-        // Check that the number string contains only digits
-        if index_str.bytes().any(|c| !c.is_ascii_digit()) {
-            return Err(DerivationError::InvalidChildNumberFormat);
-        }
-
-        // Check if the number has leading 0s
-        if index_str.len() > 1 && index_str.starts_with('0') {
-            return Err(DerivationError::InvalidChildNumberFormat);
-        }
-
-        let index = U31::from_str(index_str)?;
-
-        if is_hardened {
-            ChildNumber::from_hardened(index)
-        } else {
-            ChildNumber::from_normal(index)
-        }
-    }
-}
-
-impl fmt::Display for ChildNumber {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self.0 {
-            DerivationType::Hardened(index) => {
-                fmt::Display::fmt(&index, f)?;
-                let alt = f.alternate();
-                f.write_char(if alt { HARDENED_H } else { HARDENED_APOS })
-            }
-        }
-    }
-}
 
 /// BIP-32 compatible derivation path
 #[derive(Debug, Clone, PartialEq, Eq, Ord, PartialOrd)]
